backend/seo/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
python-dotenv
wordcloud
aiofiles
certifi
cryptography
pyOpenSSL
OpenSSL
httpx[http2]
python-multipart
openai
ruff
ssl
dnspython